refactor: test about check subcommand was modularized
